Added new commands for the menu bar.
Added a command for creating a new character, opening a character, saving a character, and saving a character as. Also started preliminary work on getting some of the save functionality working.  Next steps would be to actually start setting up the models for the character. Get that on the screen, and then start worrying about wiring all the commands and such together.