Learned something new about autofac that I wanted to test.
Looks like I should be able to do some of the things that I was trying to do at application startup that I wasn't able to because the Eto.Application wasn't running yet.  I found out that I can force the Eto.Application to initialize and start running before the rest of the container is built.  This will allow me to configure stuff like NLog to work correctly without having to do it as the Eto.Application is initilizing. YAY!!